Switch quilt and quilt_rails test directories
