+++ gems/quilt_rails/test/quilt/configuration_test.rb
