+++ gems/quilt_rails/test/quilt/header_csrf_strategy_test.rb
